+++ CHANGELOG.md
@@ -28,3 +28,4 @@
 - Item Magnet will show in the tooltip that an item is blacklisted (configurable)
 - Added a recipe for unsetting Anti-Plasmids, if for whatever reason you want to
 - Plasmids and Anti-Plasmids now only stack to 1
+- Fixed the Black Death recipe being broken